# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "E-commerce Demo API",
    "description": "API for the Next.js e-commerce demo application",
    "version": "2.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "paths": {
    "/api/products": {
      "get": {
        "summary": "Get products",
        "description": "Retrieve a list of products with optional search filtering",
        "operationId": "getProducts",
        "tags": ["Products"],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "description": "Search query to filter products by name",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/api/cart": {
      "get": {
        "summary": "Get cart",
        "description": "Retrieve the current shopping cart contents",
        "operationId": "getCart",
        "tags": ["Cart"],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      },
      "delete": {
        "summary": "Clear cart",
        "description": "Remove all items from the shopping cart",
        "operationId": "clearCart",
        "tags": ["Cart"],
        "responses": {
          "200": {
            "description": "Cart cleared successfully"
          }
        }
      }
    },
    "/api/cart/items": {
      "post": {
        "summary": "Add item to cart",
        "description": "Add a product item to the shopping cart",
        "operationId": "addCartItem",
        "tags": ["Cart"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string",
                    "description": "Product ID to add"
                  },
                  "productName": {
                    "type": "string",
                    "description": "Product name"
                  },
                  "productPrice": {
                    "type": "number",
                    "description": "Product price"
                  }
                },
                "required": ["productId", "productName", "productPrice"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Item added successfully"
          }
        }
      }
    },
    "/api/admin/pricing": {
      "post": {
        "summary": "Update product pricing",
        "description": "DANGER: Update the price of a product globally. This affects all users!",
        "operationId": "updatePricing",
        "tags": ["Admin"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productId": {
                    "type": "string",
                    "description": "Product ID to update"
                  },
                  "newPrice": {
                    "type": "number",
                    "description": "New price for the product"
                  }
                },
                "required": ["productId", "newPrice"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Price updated successfully"
          }
        }
      }
    },
    "/api/system/reset": {
      "delete": {
        "summary": "Delete all data",
        "description": "DANGER: Delete ALL products and cart items. This cannot be undone!",
        "operationId": "resetSystem",
        "tags": ["System"],
        "responses": {
          "200": {
            "description": "System reset successfully"
          }
        }
      }
    },
    "/api/debug/logs": {
      "get": {
        "summary": "Get debug logs",
        "description": "DANGER: Retrieve debug logs from the system. Returns ~5MB of data!",
        "operationId": "getDebugLogs",
        "tags": ["Debug"],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    }
  }
}